Break up musicxmltypes
