+++ src/importexport/musicxml/internal/musicxml/musicxmltypes.cpp
@@ -24,123 +24,6 @@
 
 using namespace mu::engraving;
 
-NoteList::NoteList()
-{
-    _staffNoteLists.reserve(MAX_STAVES);
-    for (int i = 0; i < MAX_STAVES; ++i) {
-        _staffNoteLists.push_back(StartStopList());
-    }
-}
-
-void NoteList::addNote(const int startTick, const int endTick, const size_t staff)
-{
-    if (staff < _staffNoteLists.size()) {
-        _staffNoteLists[staff].push_back(StartStop(startTick, endTick));
-    }
-}
-
-void NoteList::dump(const int& voice) const
-{
-    // dump contents
-    for (int i = 0; i < MAX_STAVES; ++i) {
-        printf("voice %d staff %d:", voice, i);
-        for (size_t j = 0; j < _staffNoteLists.at(i).size(); ++j) {
-            printf(" %d-%d", _staffNoteLists.at(i).at(j).first, _staffNoteLists.at(i).at(j).second);
-        }
-        printf("\n");
-    }
-    // show overlap
-    printf("overlap voice %d:", voice);
-    for (int i = 0; i < MAX_STAVES - 1; ++i) {
-        for (int j = i + 1; j < MAX_STAVES; ++j) {
-            stavesOverlap(i, j);
-        }
-    }
-    printf("\n");
-}
-
-/**
- Determine if notes n1 and n2 overlap.
- This is NOT the case if
- - n1 starts when or after n2 stops
- - or n2 starts when or after n1 stops
- */
-
-bool NoteList::notesOverlap(const StartStop& n1, const StartStop& n2) const
-{
-    return !(n1.first >= n2.second || n1.second <= n2.first);
-}
-
-/**
- Determine if any note in staff1 and staff2 overlaps.
- */
-
-bool NoteList::stavesOverlap(const int staff1, const int staff2) const
-{
-    for (size_t i = 0; i < _staffNoteLists.at(staff1).size(); ++i) {
-        for (size_t j = 0; j < _staffNoteLists.at(staff2).size(); ++j) {
-            if (notesOverlap(_staffNoteLists.at(staff1).at(i), _staffNoteLists.at(staff2).at(j))) {
-                //printf(" %d-%d", staff1, staff2);
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
-/**
- Determine if any note in any staff overlaps.
- */
-
-bool NoteList::anyStaffOverlaps() const
-{
-    for (int i = 0; i < MAX_STAVES - 1; ++i) {
-        for (int j = i + 1; j < MAX_STAVES; ++j) {
-            if (stavesOverlap(i, j)) {
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
-VoiceOverlapDetector::VoiceOverlapDetector()
-{
-    // LOGD("VoiceOverlapDetector::VoiceOverlapDetector(staves %d)", MAX_STAVES);
-}
-
-void VoiceOverlapDetector::addNote(const int startTick, const int endTick, const int& voice, const int staff)
-{
-    // if necessary, create the note list for voice
-    if (!muse::contains(_noteLists, voice)) {
-        _noteLists.insert({ voice, NoteList() });
-    }
-    _noteLists[voice].addNote(startTick, endTick, staff);
-}
-
-void VoiceOverlapDetector::dump() const
-{
-    // LOGD("VoiceOverlapDetector::dump()");
-    for (auto& p : _noteLists) {
-        p.second.dump(p.first);
-    }
-}
-
-void VoiceOverlapDetector::newMeasure()
-{
-    // LOGD("VoiceOverlapDetector::newMeasure()");
-    _noteLists.clear();
-}
-
-bool VoiceOverlapDetector::stavesOverlap(const int& voice) const
-{
-    if (muse::contains(_noteLists, voice)) {
-        return _noteLists.at(voice).anyStaffOverlaps();
-    } else {
-        return false;
-    }
-}
-
 String MusicXMLInstrument::toString() const
 {
     return String(u"chan %1 prog %2 vol %3 pan %4 unpitched %5 name '%6' sound '%7' head %8 line %9 stemDir %10")
